feat(setup): add `--dry-run` flag (#492)
* add dry-fun flag for setup step

* add help message for --dry-run option

* fix: update test snapshot

+++ packages/shipjs/src/step/setup/addCircleCIConfig.js
@@ -13,6 +13,7 @@ export default ({
   scheduleCircleCI,
   cronExpr,
   dir,
+  dryRun,
 }) =>
   runStep(
     {
@@ -42,8 +43,13 @@ export default ({
         gitUserName: getGitConfig('user.name') || 'Your Name',
         gitUserEmail: getGitConfig('user.email') || 'your@email.com',
       });
-      mkdirp.sync(path.dirname(filePath));
-      fs.writeFileSync(filePath, content);
+      if (dryRun) {
+        print(`.circleci/config.yml`);
+        print(content);
+      } else {
+        mkdirp.sync(path.dirname(filePath));
+        fs.writeFileSync(filePath, content);
+      }
       return () => {
         print(`${info('✔')} Created \`.circleci/config.yml\`.`);
         print('  You still need to finish setting up at CircleCI.');
